+++ trantor/utils/MsgBuffer.cc
@@ -20,6 +20,7 @@
 #include <netinet/in.h>
 #else
 #include <WindowsSupport.h>
+#include <winsock2.h>
 #endif
 #include <errno.h>
 #include <assert.h>
@@ -150,7 +151,7 @@ ssize_t MsgBuffer::readFd(int fd, int *retErrno)
     struct iovec vec[2];
     size_t writable = writableBytes();
     vec[0].iov_base = begin() + tail_;
-    vec[0].iov_len = writable;
+    vec[0].iov_len = static_cast<int>(writable);
     vec[1].iov_base = extBuffer;
     vec[1].iov_len = sizeof(extBuffer);
     const int iovcnt = (writable < sizeof extBuffer) ? 2 : 1;
